chore: move types to separate files under types folder
