+++ lib/types/index.d.ts
@@ -1,0 +1,3 @@
+export * from './atom'
+export * from './intentions'
+export * from './linter'
